deployment/requirements.txt
streamlit
pandas
numpy
tensorflow==2.15.0
keras==2.15.0
Pillow
